requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.2
django-allauth==0.63.6
django-appconf==1.1.0
django-cors-headers==4.4.0
django-crispy-forms
crispy-bootstrap5
django-debug-toolbar==4.4.6
django-filter==24.3
django-formtools==2.5.1
django-jinja==2.11.0
django-redis==5.4.0
django-seed==0.3.1
django-select2==8.2.0
django-tables2==2.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==37.1.0
pillow==11.2.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
redis==5.0.8
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
django-summernote
xhtml2pdf==0.2.16
